Retrieve KOM settings from order_id when logging
Currently, we require the KOM settings to determine whether logging is enabled. However, the function for retrieving the settings is only available in the KOM_Request functions while logging like in the case of Klarna Sellers App is happening elsewhere as well. With this change, we'll now use the WC order ID which is more readily available to retrieve the KOM settings.